requirements.txt
Django
django-environ
gunicorn
psycopg2
requests
djangorestframework
cffi
cryptography
pathlib
pycparser
PyJWT
whitenoise
django-cors-headers
django-allauth
django-rest-auth
django-phonenumber-field
django-phonenumbers
pillow
django-filter
dj-rest-auth
pyotp
